Create Bank WD Ternary Operator
The program is bank card withdrawal system: it takes the account and the amount that the user wants to withdraw, then outputs the remaining money. If the requested cash is greater than the balance, the program outputs "Error".
The bank wants to set a minimal value of $500 for withdrawal.  The output is "Error" if the requested money is less than $500.

Sample Input
4500
300

Sample Output
Error